requirements.txt
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.2
django-js-asset==2.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
